refactor(tests): Remove redundant `clearOutput()` test and enhance output buffer handling in `StatelessApplicationTest` class.
