+++ src/http/ErrorHandler.php
@@ -72,7 +72,9 @@ final class ErrorHandler extends \yii\web\ErrorHandler
 
         while ($currentLevel > $minLevel) {
             if (@ob_end_clean() === false) {
+                // @codeCoverageIgnoreStart
                 ob_clean();
+                // @codeCoverageIgnoreEnd
             }
 
             $currentLevel = ob_get_level();
